pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragwatch"
version = "0.1.0"
description = "OpenTelemetry-native RAG observability SDK with semantic quality scores"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "RAGWatch Contributors"},
]
keywords = ["opentelemetry", "rag", "observability", "tracing", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "opentelemetry-api==1.24.0",
    "opentelemetry-sdk==1.24.0",
]

[project.optional-dependencies]
langgraph = [
    "langgraph>=0.2.0",
]
crewai = [
    "crewai>=0.41.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/ragwatch/ragwatch"
Repository = "https://github.com/ragwatch/ragwatch"
Issues = "https://github.com/ragwatch/ragwatch/issues"

[tool.hatch.build.targets.wheel]
packages = ["ragwatch"]

[tool.uv]
python-preference = "only-managed"

[tool.uv.pip]
extra-index-url = []

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
